# Faster `sparse` for `AbstractFill`