--- a/ext/FillArraysSparseArraysExt.jl
+++ b/ext/FillArraysSparseArraysExt.jl
@@ -5,6 +5,7 @@ using SparseArrays: SparseVectorUnion
 import Base: convert, kron
 using FillArrays
 using FillArrays: RectDiagonalFill, RectOrDiagonalFill, ZerosVector, ZerosMatrix, getindex_value, AbstractFillVector, _fill_dot
+using FillArrays: AbstractFill
 # Specifying the full namespace is necessary because of https://github.com/JuliaLang/julia/issues/48533
 # See https://github.com/JuliaStats/LogExpFunctions.jl/pull/63
 using FillArrays.LinearAlgebra
@@ -18,6 +19,13 @@ SparseVector{Tv,Ti}(Z::ZerosVector) where {Tv,Ti} = spzeros(Tv, Ti, length(Z))
 
 convert(::Type{AbstractSparseVector}, Z::ZerosVector{T}) where T = spzeros(T, length(Z))
 convert(::Type{AbstractSparseVector{T}}, Z::ZerosVector) where T= spzeros(T, length(Z))
+
+function SparseMatrixCSC{T,Ti}(F::AbstractFill) where {T, Ti<:Integer}
+    SparseMatrixCSC{T,Ti}(size(F)...,
+        Vector(StepRangeLen(1, size(F,1), size(F,2)+1)),
+        convert(Vector, reduce(vcat, fill(axes(F,1), size(F,2)))),
+        fill(convert(T, getindex_value(F)), length(F)))
+end
 
 SparseMatrixCSC{T}(Z::ZerosMatrix) where T = spzeros(T, size(Z)...)
 SparseMatrixCSC{Tv,Ti}(Z::Zeros{T,2,Axes}) where {Tv,Ti<:Integer,T,Axes} = spzeros(Tv, Ti, size(Z)...)