--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -592,7 +592,13 @@ end
         end
     end
 
-    @test sparse(Fill(3, 4, 4)) == sparse(fill(3, 4, 4))
+    F = Fill(3, 4, 4)
+    @test sparse(F) == sparse(fill(3, size(F)...))
+    @test SparseMatrixCSC{Int8,Int16}(F) isa SparseMatrixCSC{Int8,Int16}
+
+    F = Fill(3, 4)
+    @test sparse(F) == sparse(fill(3, size(F)...))
+    @test SparseVector{Int8,Int16}(F) isa SparseVector{Int8,Int16}
 end
 
 @testset "==" begin