There aren't particularly useful, but can't hurt to have some faster methods.
```julia
julia> F = Fill(3, 4_000);

julia> @b sparse(F)
11.541 μs (19 allocs: 197.219 KiB) # master
3.501 μs (7 allocs: 62.656 KiB) # This PR

julia> F = Fill(3, 1_000, 1_000);

julia> @b SparseMatrixCSC{Int32,Int32}(F)
5.623 ms (52 allocs: 17.417 MiB) # master
546.454 μs (19 allocs: 7.653 MiB) # This PR
```